improve how the path to the custom rules is resolved
When used in combination of a git-hook process.env.PWD returns the path to the folder where the script has been initiated.
This results in the inability to run the git-hook from any other directory.

For example git push (which triggers the git-hook) can only be executed from the "root" folder of the project.

When we change this to process.cwd() we will always get the folder where the Gruntfile.js is located.
This makes it possible to then trigger the git-hook in any folder of the project.